Solve two int-size issues on some platforms; exit is well defined to return an (int) while on some platforms read/write is still expressed as unsigned int bytes of data.  Both harmless truncations.
git-svn-id: https://svn.apache.org/repos/asf/apr/apr/trunk@584587 13f79535-47bb-0310-9956-ffa450edef68

+++ test/sockchild.c
@@ -67,14 +67,14 @@ int main(int argc, char *argv[])
             exit(-1);
         }
         
-        exit(length);
+        exit((int)length);
     }
     else if (!strcmp("write", argv[1])) {
         apr_size_t length = strlen(DATASTR);
         apr_socket_send(sock, DATASTR, &length);
 
         apr_socket_close(sock);
-        exit(length);
+        exit((int)length);
     }
     exit(-1);
 }
